fix: refactored, lang list
